[shell] enable shell as an optional module in apps
This PR allows users to build shell into the devices for debugging and
testing purpose.
