+++ examples/all-clusters-app/esp32/sdkconfig.defaults
@@ -45,3 +45,6 @@ CONFIG_DEVICE_PRODUCT_ID=0x4541
 # Main task needs a bit more stack than the default
 # default is 3584, bump this up to 4k.
 CONFIG_ESP_MAIN_TASK_STACK_SIZE=4096
+
+#enable debug shell
+CONFIG_ENABLE_CHIP_SHELL=y
