+++ examples/platform/linux/BUILD.gn
@@ -19,6 +19,10 @@ config("app-main-config") {
   include_dirs = [ "." ]
 }
 
+declare_args() {
+  chip_enable_shell = false
+}
+
 source_set("app-main") {
   sources = [
     "AppMain.cpp",
@@ -27,14 +31,21 @@ source_set("app-main") {
     "Options.h",
   ]
 
+  defines = []
+
   if (chip_enable_pw_rpc) {
-    defines = [ "PW_RPC_ENABLED" ]
+    defines += [ "PW_RPC_ENABLED" ]
   }
 
   public_deps = [
     "${chip_root}/src/app/server",
     "${chip_root}/src/lib",
   ]
+
+  if (chip_enable_shell) {
+    defines += [ "CHIP_ENABLE_SHELL" ]
+    public_deps += [ "${chip_root}/src/lib/shell" ]
+  }
 
   public_configs = [ ":app-main-config" ]
 }
