+++ src/lib/shell/tests/TestStreamerStdio.cpp
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-#include "TestShell.h"
+#include "TestStreamerStdio.h"
 
 #include <shell/Engine.h>
 #include <support/CodeUtils.h>
